fix #19 and actually commit fix this time :/

+++ src/dsp/sample.hxx
@@ -58,7 +58,7 @@ public:
 
 	/// returns the buffer for the provided channel
 	const float*  getAudio(int channel);
-	const int     getStartPoint() { return startPoint; }
+	const float   getStartPoint() { return startPoint; }
 
 	/// returns the waveform buffer, a mono-mixdown resampled to fit the window
 	const float* getWaveform();
